Plot misfit map
plot_misfit_map.py will plot misfit-related metrics on a map. It relies on the presence of LASIF (www.lasif.net, Krischer et al, Seismological Research Letters 2015, doi:10.1785/0220140248) on the system and the lasif environment activated ($ source activate lasif).

plot_misfit_map contains the following functions:
- plot_title_legend    -- only relevant on a plot that already exists: just didn't want to bother with the formatting in the main code
- plot_coloured_hist   -- very nice trick (from stackoverflow, actually, but I don't remember where, sorry!) to plot a histogram in which the bars also have a colour according to a given colour scale.
- plot_misfits_map     -- plots any type of misfit-related number on a map, provided that you've already calculated the lats and lons in the right projection. You can choose to also plot the stations for which no windows were selected, give the colour range, plot the earthquake focal mechanism ('beachball') and plot a histogram or not
- plot_misfit_map_avg  -- Does the pre calculations for plot_misfit_map, in this case calculating the total misfit per station or the average phase difference integral (closely related to the average phase shift, but not quite it)
- plot_misfit_change_map -- similar to above, but will now plot CHANGES in misfit or phaseshift integral (or future: other visualisation metrics) from one iteration to another.

IMPORTANT: the misfit map plots rely on a change to the lasif code itself, namely in LASIF/lasif/adjoint_sources/ad_src_tf_phase_misfit.py (see this repository, but you'll have to add it to your personal lasif installation in the correct location.

IMPORTANT: in this version of the code, the phasediff_integral is actually assumed to be saved in a 'mean_delay' key in the window details dictionary. This is nonsense, so I will rename it to congruent terms